Move settings types and initial values to lib/settings
